feat: 缓存统一为 SQLite + 修复 Docker 写入权限 + L4 信号列排序
- kline_cache: 新增 financial_map 表，统一缓存到 wyckoff.db
- data_source: fetch_financial_map 改用 SQLite；旧 JSON 自动迁移
- dashboard: 启动时 chmod data/ 目录确保 Docker volume 可写
- local_db: get_db() 运行时读取 WYCKOFF_DB_PATH 环境变量
- funnel.tsx: L4 信号列支持点击排序
- sector_rotation: 修复 pandas FutureWarning 刷屏

+++ core/sector_rotation.py
@@ -96,7 +96,7 @@ def _member_snapshot(df: pd.DataFrame) -> dict | None:
     pct = close.pct_change() * 100.0
     prev_close = close.shift(1)
     amplitude = (high - low) / prev_close.replace(0, pd.NA) * 100.0
-    close_pos = ((close - low) / (high - low).replace(0, pd.NA) * 100.0).clip(0, 100).fillna(50.0)
+    close_pos = ((close - low) / (high - low).replace(0, pd.NA) * 100.0).clip(0, 100).astype(float).fillna(50.0)
 
     recent = pd.DataFrame(
         {
